refactor: table 의 package 구조를 수정하라
