+++ src/main/java/kitchenpos/table/dto/request/OrderTableRequest.java
@@ -1,27 +1,17 @@
-package kitchenpos.table.domain.response;
+package kitchenpos.table.dto.request;
 
-import kitchenpos.table.domain.OrderTable;
 
-public class OrderTableResponse {
-
+public class OrderTableRequest {
     private Long id;
     private Long tableGroupId;
     private int numberOfGuests;
     private boolean empty;
 
-    private OrderTableResponse(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
+    public OrderTableRequest(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
         this.id = id;
         this.tableGroupId = tableGroupId;
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
-    }
-
-    public static OrderTableResponse of(OrderTable orderTable) {
-        return new OrderTableResponse(
-            orderTable.getId()
-            , orderTable.getTableGroupId()
-            , orderTable.getNumberOfGuests()
-            , orderTable.isEmpty());
     }
 
     public Long getId() {
