+++ src/main/java/kitchenpos/table/dto/response/OrderTableResponse.java
@@ -1,17 +1,27 @@
-package kitchenpos.table.domain.request;
+package kitchenpos.table.dto.response;
 
+import kitchenpos.table.domain.OrderTable;
 
-public class OrderTableRequest {
+public class OrderTableResponse {
+
     private Long id;
     private Long tableGroupId;
     private int numberOfGuests;
     private boolean empty;
 
-    public OrderTableRequest(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
+    private OrderTableResponse(Long id, Long tableGroupId, int numberOfGuests, boolean empty) {
         this.id = id;
         this.tableGroupId = tableGroupId;
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
+    }
+
+    public static OrderTableResponse of(OrderTable orderTable) {
+        return new OrderTableResponse(
+            orderTable.getId()
+            , orderTable.getTableGroupId()
+            , orderTable.getNumberOfGuests()
+            , orderTable.isEmpty());
     }
 
     public Long getId() {
