+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -15,8 +15,8 @@ import kitchenpos.menu.domain.response.MenuResponse;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.request.OrderRequest;
 import kitchenpos.order.domain.response.OrderResponse;
-import kitchenpos.table.domain.request.OrderLineItemRequest;
-import kitchenpos.table.domain.response.OrderTableResponse;
+import kitchenpos.order.domain.request.OrderLineItemRequest;
+import kitchenpos.table.dto.response.OrderTableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
