+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -9,9 +9,9 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.table.domain.request.OrderTableRequest;
+import kitchenpos.table.dto.request.OrderTableRequest;
 import kitchenpos.tableGroup.dto.request.TableGroupRequest;
-import kitchenpos.table.domain.response.OrderTableResponse;
+import kitchenpos.table.dto.response.OrderTableResponse;
 import kitchenpos.tableGroup.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
