+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -28,9 +28,9 @@ import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.tableGroup.application.TableGroupService;
 import kitchenpos.tableGroup.domain.TableGroup;
 import kitchenpos.tableGroup.domain.TableGroupRepository;
-import kitchenpos.table.domain.request.OrderTableRequest;
+import kitchenpos.table.dto.request.OrderTableRequest;
 import kitchenpos.tableGroup.dto.request.TableGroupRequest;
-import kitchenpos.table.domain.response.OrderTableResponse;
+import kitchenpos.table.dto.response.OrderTableResponse;
 import kitchenpos.tableGroup.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
