+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -22,8 +22,8 @@ import kitchenpos.table.application.OrderTableServiceImpl;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.table.domain.request.OrderTableRequest;
-import kitchenpos.table.domain.response.OrderTableResponse;
+import kitchenpos.table.dto.request.OrderTableRequest;
+import kitchenpos.table.dto.response.OrderTableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
